+++ config/deploy/iacseaturtle-ort/staging.rb
@@ -1,6 +1,6 @@
 set :stage, :staging
 
-set :branch, :master
+set :branch, 'master'
 
 server "ort-staging.linode.unep-wcmc.org", user: "wcmc", roles: %w{app web db}